Delete all command classes. Change identifier type(int -> long).

+++ expense-tracker-server/src/main/java/inzagher/expense/tracker/server/model/criteria/ExpenseSearchCriteria.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 public class ExpenseSearchCriteria implements Serializable {
-    private final List<Integer> categoryIdentifiers = new ArrayList<>();
-    private final List<Integer> personIdentifiers = new ArrayList<>();
+    private final List<Long> categoryIdentifiers = new ArrayList<>();
+    private final List<Long> personIdentifiers = new ArrayList<>();
     private BigDecimal amountExact;
     private BigDecimal amountFrom;
     private BigDecimal amountTo;
